refactor(stella-tools): split process.rs's tests into process/tests.rs
clear_output pushed process.rs to 1598 lines, over the file-size gate's
1500-line limit — it is not grandfathered (stella-tools's only god files
are registry.rs and scripts.rs, per AGENTS.md), so this is a first-time
crossing that gets split, not a baseline entry. Moves the #[cfg(test)] mod
tests block into process/tests.rs, mirroring the existing registry.rs /
registry/tests.rs pattern in the same crate. No behavior change; every test
moved verbatim (799 stella-tools tests still pass) plus a fixed unused-var
name in the moved witness test.

Also fixes two now-stale generated counts docs_in_sync.rs checks against
catalog.rs (agent-tools/index.mdx's "N tools register in every session" and
"All told: N always-on tools, up to M native"), both one-off from
clear_output joining the always-on/native tool set.
